Move functions chapter components into dedicated local component folders
