+++ content/functions/components/card-graph/card-graph.scss
