+++ content/functions/components/draw-graph/draw-graph.scss
