+++ content/functions/components/function-machine/function-machine.less
