+++ content/functions/components/function-machine/function-machine.scss
