+++ content/functions/components/video-graph/video-graph.scss
